--- app/src/main/java/com/ikseong/ucompass/data/socket/repository/UserLocation.kt
@@ -1,3 +1,0 @@
-package com.ikseong.ucompass.data.socket.repository
-
-data class UserLocation(var lat:Double , var lng:Double)
